# Backend_Xsigma/api/openapi.yaml
openapi: 3.0.3
info:
  title: Analytical Sigma Volatility API
  description: |
    Advanced quantitative finance API for volatility modeling and risk analytics.

    This API provides comprehensive volatility surface calculations, FX volatility modeling,
    and advanced risk analytics using state-of-the-art mathematical frameworks including
    Extended SVI models and quadratic smile interpolation.

    ## Features
    - Extended SVI volatility surface modeling
    - FX volatility curve interpolation
    - Multi-model volatility comparison
    - Real-time market data integration
    - Advanced caching and performance optimization
    - Comprehensive error handling and validation

    ## Models Supported
    - Analytical Sigma Volatility (Extended SVI)
    - FX Volatility Models (Call/Put, Instrument, Delta, SVI)
    - Volatility Surface Generation
    - Probability Density Functions
    - Risk Analytics

  version: 2.1.0
  contact:
    name: XSigma Team
    email: support@xsigma.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5005
    description: Development server
  - url: https://api.xsigma.com
    description: Production server
...
paths:
...
  /api/analytical-sigma:
    get:
      tags:
        - Analytical Sigma
      summary: Calculate Analytical Sigma Volatility
      description: |
        Calculate volatility surfaces using the Extended SVI model with 4 test cases.
        
        **Test Cases:**
        - `test=1`: Basic volatility surface with default parameters for large forward values
        - `test=2`: Sensitivity analysis comparing different ctrl_c values
        - `test=3`: Probability density calculation with analytical and bump methods
        - `test=4`: Cumulative probability analysis with monotonicity checks
        
        **Performance:** Results are cached for 5 minutes to improve response times.
      parameters:
        - name: test
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 4
            default: 1
          description: |
            Test case number:
            - 1: volatility_surface (default)
            - 2: vols_plus_minus 
            - 3: density
            - 4: probability
          example: 1
        - name: n
          in: query
          required: false
          schema:
            type: integer
            minimum: 10
            maximum: 2000
            default: 200
          description: Number of data points to calculate
          example: 200
        - name: fwd
          in: query
          required: false
          schema:
            type: number
            minimum: 0.001
          description: |
            Forward price. Defaults depend on test case:
            - Test 1: 2245.07 (large forward scenario)
            - Tests 2-4: 1.0 (normalized scenario)
          example: 2245.0656707892695
        - name: time
          in: query
          required: false
          schema:
            type: number
            minimum: 0.001
          description: |
            Time to expiration. Defaults depend on test case:
            - Test 1: 1.0 (1 year)
            - Tests 2-4: 0.333 (4 months)
          example: 1.0
        - name: ctrl_p
          in: query
          required: false
          schema:
            type: number
            default: 0.2
          description: Control parameter P for SVI model
          example: 0.2
        - name: ctrl_c
          in: query
          required: false
          schema:
            type: number
            default: 0.2
          description: Control parameter C for SVI model (modified in test 2)
          example: 0.2
        - name: atm
          in: query
          required: false
          schema:
            type: number
            minimum: 0.001
          description: |
            At-the-money volatility. Defaults depend on test case:
            - Test 1: 1.1 (high volatility scenario)
            - Tests 2-4: 0.1929 (low volatility scenario)
          example: 1.1
        - name: skew
          in: query
          required: false
          schema:
            type: number
          description: |
            Volatility skew parameter. Defaults depend on test case:
            - Test 1: 3.5 (high skew)
            - Tests 2-4: 0.02268 (low skew)
          example: 3.5
        - name: smile
          in: query
          required: false
          schema:
            type: number
          description: |
            Volatility smile parameter. Defaults depend on test case:
            - Test 1: 17 (pronounced smile)
            - Tests 2-4: 0.00317 (subtle smile)
          example: 17
        - name: put
          in: query
          required: false
          schema:
            type: number
          description: |
            Put wing parameter. Defaults depend on test case:
            - Test 1: 0.7 (positive)
            - Tests 2-4: -0.00213 (negative)
          example: 0.7
        - name: call
          in: query
          required: false
          schema:
            type: number
          description: |
            Call wing parameter. Defaults depend on test case:
            - Test 1: 0.06 (positive)
            - Tests 2-4: -0.00006 (negative)
          example: 0.06
      responses:
        '200':
          description: Successful calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolatilityResponse'
              examples:
                test1_volatility_surface:
                  summary: Test 1 - Basic volatility surface
                  value:
                    status: "success"
                    data:
                      strikes: [561.26, 573.45, 585.64]
                      Tab_1: [1.2345, 1.2456, 1.2567]
                      Tab_2: [1.1000, 1.1000, 1.1000]
                      metadata:
                        test_case: 1
                        description: "Basic volatility surface"
                        execution_time_ms: 2.01
                        data_points: 200
                    api:
                      version: "2.0.0"
                      testCase:
                        number: 1
                        name: "volatility_surface"
                        description: "Basic volatility surface calculation"
                      responseTime: 15
                      cached: false
                test2_sensitivity:
                  summary: Test 2 - Sensitivity analysis
                  value:
                    status: "success"
                    data:
                      strikes: [0.3, 0.31, 0.32]
                      Tab_1: [0.162, 0.161, 0.161]
                      Tab_2: [0.162, 0.161, 0.161]
                      metadata:
                        test_case: 2
                        description: "Sensitivity analysis (ctrl_c variations)"
                        execution_time_ms: 2.00
                        ctrl_c_original: 0.2
                        ctrl_c_modified: 4.0
                        sensitivity_metrics:
                          average_difference: 0.0178
                          maximum_difference: 0.0681
                    api:
                      version: "2.0.0"
                      testCase:
                        number: 2
                        name: "vols_plus_minus"
                        description: "Sensitivity analysis with ctrl_c variations"
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Volatility
      summary: Calculate volatility surface (POST)
      description: Same functionality as GET endpoint but with request body for complex parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolatilityRequest'
            examples:
              test1_basic:
                summary: Test 1 - Basic volatility surface with defaults
                value:
                  test: 1
                  n: 200
              test2_sensitivity:
                summary: Test 2 - Sensitivity analysis with custom parameters
                value:
                  test: 2
                  n: 100
                  fwd: 1.0
                  time: 0.333
                  atm: 0.1929
                  skew: 0.02268
              test3_density:
                summary: Test 3 - Density analysis
                value:
                  test: 3
                  n: 200
                  fwd: 1.0
                  time: 0.333
                  ctrl_p: 0.2
                  ctrl_c: 0.2
                  atm: 0.1929
                  skew: 0.02268
                  smile: 0.00317
                  put: -0.00213
                  call: -0.00006
              test4_probability:
                summary: Test 4 - Probability analysis
                value:
                  test: 4
                  n: 150
                  fwd: 1.0
                  time: 0.333
                  atm: 0.1929
                  skew: 0.02268
      responses:
        '200':
          description: Successful calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolatilityResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/fx-volatility/atm-curve:
    get:
      summary: Get ATM Volatility Curve
      description: Retrieve interpolated ATM volatility curve with market data points
      tags: [FX Volatility]
      parameters:
        - name: refresh
          in: query
          description: Force refresh cache
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: ATM volatility curve data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FXVolatilityATMResponse'
  /api/fx-volatility/models-comparison:
    get:
      summary: Compare FX Volatility Models
      description: Compare Call/Put, Instrument, Delta, and SVI volatility models
      tags: [FX Volatility]
      parameters:
        - name: expiry
          in: query
          description: Expiry time for comparison
          schema:
            type: number
            minimum: 0.01
            maximum: 10
        - name: refresh
          in: query
          description: Force refresh cache
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Model comparison data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FXVolatilityModelsResponse'
...
  /api/AnalyticalSigmaVolatilityCalibration:
    post:
      tags:
        - Analytical Sigma
      summary: Calibrate volatility models and compute volatility surfaces or density functions
      description: |
        Calibrates volatility models to market data and computes volatility surfaces or probability density functions.

        ## Supported Computation Types
        - **volatility_asv**: Analytical Sigma Volatility model calibration with implied volatilities
        - **density**: Computes probability density functions based on calibrated model
        - **volatility_svi**: Stochastic Volatility Inspired model calibration
        - **dynamic_asv**: Interactive ASV with parameter sliders
        - **dynamic_svi**: Interactive SVI with parameter sliders

        ## Features
        - Model calibration using Ceres solver
        - Interactive parameter adjustment with sliders
        - Real-time volatility smile and density plotting
        - 5 graphs: 3 stable (calibration data) + 2 dynamic (interactive)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalibrationRequest'
      responses:
        '200':
          description: Calibration results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalibrationResponse'
components:
  schemas:
...
    FXVolatilityATMResponse:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
          properties:
            expiries:
              type: array
              items:
                type: number
            interpolated_vols:
              type: array
              items:
                type: number
            market_expiries:
              type: array
              items:
                type: number
            market_vols:
              type: array
              items:
                type: number
    FXVolatilityModelsResponse:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
          properties:
            strike_ratio:
              type: array
              items:
                type: number
            vol_call_put:
              type: array
              items:
                type: number
            vol_instrument:
              type: array
              items:
                type: number
            vol_delta:
              type: array
              items:
                type: number
            vol_svi:
              type: array
              items:
                type: number
...
    CalibrationRequest:
      type: object
      required:
        - computationType
        - n
        - spot
        - expiry
        - r
        - q
        - beta
        - rho
        - volvol
      properties:
        computationType:
          type: string
          enum: [volatility_asv, density, volatility_svi, dynamic_asv, dynamic_svi]
          description: |
            Type of computation to perform:
            - volatility_asv: Analytical Sigma Volatility model calibration
            - density: Probability density function calculation
            - volatility_svi: Stochastic Volatility Inspired model
            - dynamic_asv: Interactive ASV with parameter sliders
            - dynamic_svi: Interactive SVI with parameter sliders
          example: volatility_asv
        n:
          type: integer
          minimum: 10
          maximum: 2000
          description: Number of data points to calculate
          example: 200
        spot:
          type: number
          minimum: 0.001
          description: Current spot price
          example: 2245.0656
        expiry:
          type: number
          minimum: 0.001
          description: Time to expiry in years
          example: 1.0
        r:
          type: number
          description: Risk-free interest rate
          example: 0.003
        q:
          type: number
          description: Dividend yield
          example: 0.0022
        beta:
          type: number
          minimum: 0
          maximum: 1
          description: Beta parameter for SABR model (0 = normal, 1 = lognormal)
          example: 0.4158
        rho:
          type: number
          minimum: -1
          maximum: 1
          description: Correlation between asset and volatility
          example: 0.2256
        volvol:
          type: number
          minimum: 0.001
          description: Volatility of volatility
          example: 0.2256
        # Dynamic parameters for interactive models
        fwd:
          type: number
          description: Forward price for dynamic models
          example: 1.0
        time:
          type: number
          description: Time to expiry for dynamic models
          example: 0.333
        ctrl_p:
          type: number
          description: Control parameter P for ASV model
          example: 0.2
        ctrl_c:
          type: number
          description: Control parameter C for ASV model
          example: 0.2
        atm:
          type: number
          description: At-the-money volatility for ASV model
          example: 0.1929
        skew:
          type: number
          description: Volatility skew parameter for ASV model
          example: 0.02268
        smile:
          type: number
          description: Volatility smile parameter for ASV model
          example: 0.00317
        put:
          type: number
          description: Put wing parameter for ASV model
          example: 0.00213
        call:
          type: number
          description: Call wing parameter for ASV model
          example: 0.00006
        # SVI specific parameters
        b:
          type: number
          description: SVI parameter b
          example: 0.1
        m:
          type: number
          description: SVI parameter m
          example: 0.01
        sigma:
          type: number
          description: SVI parameter sigma
          example: 0.4
      example:
        computationType: "volatility_asv"
        n: 200
        spot: 2245.0656
        expiry: 1.0
        r: 0.003
        q: 0.0022
        beta: 0.4158
        rho: 0.2256
        volvol: 0.2256
    CalibrationResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        computationType:
          type: string
          enum: [volatility_asv, density, volatility_svi]
        data:
          type: object
          properties:
            calibration_strikes:
              type: array
              items:
                type: number
              description: Strike prices used for calibration
            bid_values:
              type: array
              items:
                type: number
              description: Bid volatility values
            ask_values:
              type: array
              items:
                type: number
              description: Ask volatility values
            mid_values:
              type: array
              items:
                type: number
              description: Mid volatility values
            strikes:
              type: array
              items:
                type: number
              description: Strike prices for computed values
            vols:
              type: array
              items:
                type: number
              description: Computed volatility values (for volatility computations)
            density:
              type: array
              items:
                type: number
              description: Probability density values (for density computations)
        performance:
          type: object
          properties:
            execution_time_ms:
              type: number
              description: Execution time in milliseconds
        metadata:
          type: object
          properties:
            execution_time_ms:
              type: number
            timestamp:
              type: string
              format: date-time
            computation_type:
              type: string
            cached:
              type: boolean
            test_case:
              type: string
        error:
          type: string
          nullable: true
      required:
        - status
    VolatilityRequest:
      type: object
      properties:
        test:
          type: integer
          minimum: 1
          maximum: 4
          default: 1
          description: Test case number
        n:
          type: integer
          minimum: 10
          maximum: 2000
          default: 200
          description: Number of data points
        fwd:
          type: number
          minimum: 0.001
          description: Forward price
        time:
          type: number
          minimum: 0.001
          description: Time to expiration
        ctrl_p:
          type: number
          default: 0.2
          description: Control parameter P
        ctrl_c:
          type: number
          default: 0.2
          description: Control parameter C
        atm:
          type: number
          minimum: 0.001
          description: At-the-money volatility
        skew:
          type: number
          description: Volatility skew
        smile:
          type: number
          description: Volatility smile
        put:
          type: number
          description: Put wing parameter
        call:
          type: number
          description: Call wing parameter
      example:
        test: 1
        n: 200
        fwd: 2245.07
        time: 1.0
        ctrl_p: 0.2
        ctrl_c: 0.2
        atm: 1.1
        skew: 3.5
        smile: 17
        put: 0.7
        call: 0.06
    VolatilityResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        data:
          type: object
          properties:
            strikes:
              type: array
              items:
                type: number
              description: Strike prices array
            Tab_1:
              type: array
              items:
                type: number
              description: Primary calculation results
            Tab_2:
              type: array
              items:
                type: number
              description: Secondary calculation results
            metadata:
              type: object
              description: Calculation metadata and analytics
        error:
          type: string
          nullable: true
        performance:
          type: object
          properties:
            total_execution_time_ms:
              type: number
            python_overhead_ms:
              type: number
            memory_efficient:
              type: boolean
            cache_hits:
              type: integer
            optimization_level:
              type: string
        api:
          type: object
          properties:
            version:
              type: string
            testCase:
              type: object
              properties:
                number:
                  type: integer
                name:
                  type: string
                description:
                  type: string
            parameters:
              type: object
            responseTime:
              type: number
            cached:
              type: boolean
            cacheKey:
              type: string
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Validation error details
        timestamp:
          type: string
          format: date-time
        responseTime:
          type: number
      required:
        - status
        - error